moved regexHelperDialog to LogExpert.UI
