+++ src/LogExpert/Dialogs/ToolArgsDialog.cs
@@ -1,12 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Text;
-using System.Windows.Forms;
-using LogExpert.Classes;
+﻿using LogExpert.Classes;
 using LogExpert.Controls.LogTabWindow;
+using LogExpert.UI.Dialogs;
+
+using System;
+using System.Drawing;
+using System.Windows.Forms;
 
 
 namespace LogExpert.Dialogs
