+++ src/Logexpert.UI/Dialogs/AboutBox.cs
@@ -2,7 +2,7 @@
 using System.Reflection;
 using System.Runtime.Versioning;
 
-namespace LogExpert.Dialogs
+namespace LogExpert.UI.Dialogs
 {
     [SupportedOSPlatform("windows")]
     public partial class AboutBox : Form
